Remove the trust anchor libary code
