+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,5 +1,6 @@
+#Mon Jan 28 16:03:30 GMT 2019
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
 zipStoreBase=GRADLE_USER_HOME
 zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-4.10-bin.zip
+distributionUrl=https\://services.gradle.org/distributions/gradle-5.1.1-all.zip
