--- src/main/java/uk/gov/ida/eidas/trustanchor/InvalidTrustAnchorException.java
@@ -1,13 +1,0 @@
-package uk.gov.ida.eidas.trustanchor;
-
-public class InvalidTrustAnchorException extends RuntimeException {
-  private static final long serialVersionUID = 1L;
-
-  public InvalidTrustAnchorException(String message) {
-    super(message);
-  }
-
-  public InvalidTrustAnchorException(String message, Throwable cause) {
-    super(message, cause);
-  }
-}
