+++ src/main/java/uk/gov/ida/eidas/trustanchor/cli/metadata/ProxyNodeMetadataSigningApplication.java
@@ -1,7 +1,8 @@
-package uk.gov.ida.eidas.cli.metadata;
+package uk.gov.ida.eidas.trustanchor.cli.metadata;
 
 import picocli.CommandLine;
 import picocli.CommandLine.Command;
+import uk.gov.ida.eidas.trustanchor.cli.trustanchor.SignWithFile;
 
 @Command(name="proxy-node-metadata", description="Signs Proxy Node Metadata", subcommands={
     SignWithFile.class,
