+++ src/main/java/uk/gov/ida/eidas/trustanchor/cli/metadata/SignWithFile.java
@@ -1,16 +1,20 @@
-package uk.gov.ida.eidas.cli.trustanchor;
+package uk.gov.ida.eidas.cli.metadata;
 
 import picocli.CommandLine.Command;
 import picocli.CommandLine.Option;
+import uk.gov.ida.eidas.trustanchor.cli.metadata.SignMetadata;
 import uk.gov.ida.eidas.utils.keyloader.FileKeyLoader;
 
 import java.io.File;
+import java.security.PrivateKey;
 import java.security.cert.X509Certificate;
-import java.security.interfaces.RSAPrivateKey;
 import java.util.concurrent.Callable;
 
+import static uk.gov.ida.eidas.metadata.AlgorithmType.RSA;
+
 @Command(name="sign-with-file", description="Signs the final key set with a key loaded from a file")
-public class SignWithFile extends SignTrustAnchor implements Callable<Void> {
+public class SignWithFile extends SignMetadata implements Callable<Void> {
+
   @Option(names = { "--key" }, description = "Location of the private key to use for signing", required=true)
   private File keyFile;
 
@@ -19,8 +23,10 @@ public class SignWithFile extends SignTrustAnchor implements Callable<Void> {
 
   @Override
   public Void call() throws Exception {
-    RSAPrivateKey key = FileKeyLoader.loadRSAKey(keyFile);
+    SignMetadata.initialize();
+
+    PrivateKey key = algorithm == RSA ? FileKeyLoader.loadRSAKey(keyFile) : FileKeyLoader.loadECKey(keyFile);
     X509Certificate x509Certificate = FileKeyLoader.loadCert(certificateFile);
-    return build(key, x509Certificate);
+    return build(key, x509Certificate, algorithm);
   }
 }
