+++ src/main/java/uk/gov/ida/eidas/trustanchor/cli/trustanchor/Import.java
@@ -1,4 +1,4 @@
-package uk.gov.ida.eidas.cli.trustanchor;
+package uk.gov.ida.eidas.trustanchor.cli.trustanchor;
 
 import com.nimbusds.jose.jwk.JWK;
 import org.json.JSONObject;
