+++ src/main/java/uk/gov/ida/eidas/trustanchor/cli/trustanchor/SignWithSmartcard.java
@@ -1,7 +1,8 @@
-package uk.gov.ida.eidas.cli.metadata;
+package uk.gov.ida.eidas.trustanchor.cli.trustanchor;
 
 import picocli.CommandLine.Command;
 import picocli.CommandLine.Option;
+import uk.gov.ida.eidas.trustanchor.cli.trustanchor.SignTrustAnchor;
 import uk.gov.ida.eidas.utils.keyloader.PKCS11KeyLoader;
 
 import java.io.File;
@@ -10,7 +11,7 @@ import java.security.cert.X509Certificate;
 import java.util.concurrent.Callable;
 
 @Command(name="sign-with-smartcard", description="Signs the final key set with a key from a smartcard")
-public class SignWithSmartcard extends SignMetadata implements Callable<Void> {
+public class SignWithSmartcard extends SignTrustAnchor implements Callable<Void> {
   @Option(names = { "--config" }, description = "PKCS#11 configuration passed as a file.\nSee https://tinyurl.com/pkcs11config", required=true)
   private File pkcs11Config;
 
@@ -25,11 +26,9 @@ public class SignWithSmartcard extends SignMetadata implements Callable<Void> {
 
   @Override
   public Void call() throws Exception {
-    SignMetadata.initialize();
-
     PKCS11KeyLoader keyLoader = new PKCS11KeyLoader(sun.security.pkcs11.SunPKCS11.class, pkcs11Config, password);
     PrivateKey key = keyLoader.getSigningKey(keyAlias);
     X509Certificate certificate = keyLoader.getPublicCertificate(certAlias);
-    return build(key, certificate, algorithm);
+    return build(key, certificate);
   }
 }
